Adding transmit and receive scripts with test message
